refactor: rename config folder
