+++ common/config/resource_keys.py
