--- common/configs/__init__.py
@@ -1,7 +1,0 @@
-from common.configs.aws import AWSServices
-from common.configs.categories import VNExpressCategories
-from common.configs.date_formats import DateFormats
-from common.configs.env import EnvVariables
-from common.configs.resource_keys import ResourceKeys
-from common.configs.selectors import HTMLSelectors, VNExpressSelectors
-from common.configs.url import VNEXPRESS_CATEGORY_URL, VNEXPRESS_COVID19_URL
