+++ common/utils/s3.py
@@ -5,7 +5,7 @@ import boto3
 from dataclasses_json import DataClassJsonMixin
 from smart_open import open as s_open
 
-from common.configs.aws import AWSServices
+from common.config.aws import AWSServices
 
 
 def upload_file_s3(filename: str, bucket: str, object_name=None):
